# docs/swagger.yaml
definitions:
  application.AddBalanceRequest:
    properties:
      amount:
        type: number
    required:
    - amount
    type: object
  application.LoginRequest:
    description: Estructura del cuerpo de la solicitud para el endpoint de login.
    properties:
      password:
        description: La contraseña debe ser obligatoria.
        type: string
      username:
        description: El nombre de usuario debe ser obligatorio.
        type: string
    required:
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /account/balance/add:
    post:
      consumes:
      - application/json
      description: Permite añadir saldo en USD al balance del usuario autenticado.
        Requiere un token JWT válido.
      parameters:
      - default: Bearer <token>
        description: Token JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Cuerpo de la solicitud para añadir saldo, con el monto en USD
        in: body
        name: AddBalanceRequest
        required: true
        schema:
          $ref: '#/definitions/application.AddBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Respuesta exitosa con mensaje y balance actualizado
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Error de validación: La cantidad ingresada es inválida'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Error de usuario: Usuario no encontrado'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error interno: No se pudo actualizar la información del usuario'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Añadir saldo al usuario
      tags:
      - Account
  /auth/login:
    post:
      consumes:
      - application/json
      description: Permite a los usuarios autenticarse y obtener un token JWT para
        acceder a los endpoints protegidos.
      parameters:
      - description: Credenciales del usuario para autenticación
        in: body
        name: LoginRequest
        required: true
        schema:
          $ref: '#/definitions/application.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token JWT generado para el usuario autenticado
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error en la validación de datos enviados
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Credenciales inválidas o usuario no encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error interno al generar el token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Autenticación de usuarios
      tags:
      - Auth
swagger: "2.0"